requirements.txt
alembic==1.13.1
attrs==20.3.0
autopep8==1.5.5
blinker==1.7
certifi==2020.12.5
chardet==4.0.0
click==8.1.7
Flask==3.0.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.1.4
MarkupSafe==2.1.5
packaging==20.9
pluggy==1.4.0
psycopg2-binary==2.9.9
py==1.11.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==8.0.0
python-dateutil==2.8.1
python-dotenv==1.0.1
python-editor==1.0.4
requests==2.25.1
six==1.15.0
SQLAlchemy==2.0.25
toml==0.10.2
urllib3==1.26.4
Werkzeug==3.0.1